bite
Now instead of using "pti truc" we are going to use "bite". "bite" is as very interesting word. It has enourmous power and strength. The AHIM (Association des Hommes les plus Intelligents du Monde) recognised this word as the most complexe and mystical word that exists. Using this word you can either make laugh people or cure cancer. It's power is infinite and we still have a lot to learn. So using this word will be a great aid to our project.